add changes for netapp

+++ deploy_ceph.py
@@ -8,7 +8,7 @@ import yaml
 from fabric.api import run, sudo, execute, task, get, put, local, env
 from fabric.api import parallel
 from fabric.context_managers import hide
-from fabric.contrib.files import append
+from fabric.contrib.files import append, exists
 
 
 ceph_config_file_templ_path = os.path.join(os.path.dirname(__file__),
@@ -19,6 +19,8 @@ ceph_config_file_templ = open(ceph_config_file_templ_path).read()
 
 
 def prepare_node(release):
+    if exists("/etc/redhat-release"):
+        return
     add_ceph_dev_repo_keys = "wget -q -O- " + \
                              "'https://ceph.com/git/?p=ceph.git" + \
                              ";a=blob_plain;f=keys/release.asc' " + \
@@ -58,7 +60,7 @@ def deploy_first_mon(conf_path=deployment_config):
     params.fsid_uuid = str(uuid.uuid4())
     params.mon_ip = env.host_string
 
-    # prepare_node(params.ceph_release)
+    prepare_node(params.ceph_release)
 
     ceph_config_file = ceph_config_file_templ.format(params)
 
@@ -192,7 +194,7 @@ def add_new_osd(mon_ip, conf_path=deployment_config):
         setattr(params, attr, val)
 
     # executed on osd host
-    # prepare_node(params.ceph_release)
+    prepare_node(params.ceph_release)
 
     # put ceph config
     osd_config_file = ceph_config_file_templ.format(params)
@@ -234,6 +236,79 @@ def add_new_osd(mon_ip, conf_path=deployment_config):
 
     for cmd in prepare_cmds(commands_templ.format(params)):
         run(cmd)
+
+
+@task
+@parallel
+def netapp_add_new_osd(mon_ip, conf_path=deployment_config):
+    params = get_config(conf_path)
+    params.mon_ip = mon_ip
+
+    # executed on mon host
+
+    osd_devs = run("ls -1 /dev/sd*").split()
+    # select HDD devices
+    osd_devs = [dev for dev in osd_devs if len(os.path.nasename(dev)) == 4]
+
+    # executed on osd host
+    prepare_node(params.ceph_release)
+
+    # remove journal section
+    cfg_templ = re.sub(r"(?ims)^osd journal = .*$", "", ceph_config_file_templ)
+
+    # put ceph config
+    osd_config_file = cfg_templ.format(params)
+
+    put(remote_path=params.ceph_cfg_path,
+        local_path=StringIO(osd_config_file),
+        use_sudo=True)
+
+    # put admin keyring
+    put(remote_path=params.admin_keyring_path,
+        local_path=StringIO(params.admin_keyring_content),
+        use_sudo=True)
+
+    commands_templ = """
+    sudo mkdir -p {0.data_mount_path}
+
+    sudo mkfs -t {0.fs_type} {0.osd_data_dev}
+
+    sudo mount {0.mount_opst} {0.osd_data_dev} {0.data_mount_path}
+
+    sudo ceph-osd -c {0.ceph_cfg_path} -i {0.osd_num}
+        --cluster {0.clustername} --mkfs --mkkey --osd-uuid {0.osd_uuid}
+
+    sudo ceph auth add osd.{0.osd_num} osd 'allow *' mon 'allow profile osd' -i
+        /var/lib/ceph/osd/{0.clustername}-{0.osd_num}/keyring
+
+    ceph osd crush add-bucket {0.hostname} host
+
+    ceph osd crush move {0.hostname} root=default
+
+    sudo start ceph-osd id={0.osd_num}
+
+    ceph osd crush add {0.osd_num} {0.osd_weigth} host={0.hostname}
+
+    ceph -s
+    """
+
+    keys = {'fsid': 'fsid_uuid', 'mon initial members': 'mons'}
+
+    print ">>>>>>>>>>>>>>>>>>>", osd_devs
+
+    for dev in osd_devs:
+        params.osd_uuid = str(uuid.uuid4())
+
+        res = execute(allocate_osd_id_and_read_config, params, keys,
+                      hosts=[params.mon_ip])
+
+        for attr, val in res[params.mon_ip].items():
+            setattr(params, attr, val)
+
+        params.data_mount_path = params.data_mount_path.format(params)
+
+        for cmd in prepare_cmds(commands_templ.format(params)):
+            run(cmd)
 
 
 @task
